+++ src/rfo.cpp
@@ -24,9 +24,9 @@
 
 #include "model.h"
 
-Matrix4f RFO::GetSTLTransformationMatrix(int object, int file) const
+Matrix4d RFO::GetSTLTransformationMatrix(int object, int file) const
 {
-	Matrix4f result = transform3D.transform;
+	Matrix4d result = transform3D.transform;
 //	Vector3f translation = result.getTranslation();
 //	result.setTranslation(translation+PrintMargin);
